Now with static assets bundled in with native go!

+++ build.sh
@@ -1,22 +1,29 @@
 #! /bin/bash
 
 # kill any currently running version of the app
-killall secret-share-li
-killall secret-share-da
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+    killall secret-share-li
+fi
 
-# build the static assets so they get bundled into the go binary
-go-bindata -fs -o httpd/bindata.go views/ assets/ assets/images/
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    killall secret-share-da
+fi
 
-# cd into the httpd folder to build and run the new binary
-cd httpd
-env GOOS=linux GOARCH=amd64 go build -o ../builds/secret-share-linux
-env GOOS=darwin GOARCH=amd64 go build -o ../builds/secret-share-darwin
-env GOOS=windows GOARCH=amd64 go build -o ../builds/secret-share.exe
+cd frontend
+# npm install
+npm run build
+cd ..
+
+
+# build and run the new binary
+env GOOS=linux GOARCH=amd64 go build -o builds/secret-share-linux
+env GOOS=darwin GOARCH=amd64 go build -o builds/secret-share-darwin
+env GOOS=windows GOARCH=amd64 go build -o builds/secret-share.exe
 
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
-    ../builds/secret-share-linux &
+    builds/secret-share-linux &
 elif [[ "$OSTYPE" == "darwin"* ]]; then
-    ../builds/secret-share-darwin &
+    builds/secret-share-darwin &
 elif [[ "$OSTYPE" == "cygwin" ]]; then
     echo "windows?"
 elif [[ "$OSTYPE" == "msys" ]]; then
